Wait for the existing chat room messages to load then scroll bottom
Wrapping `$scope.scrollChat();` execution inside a `setTimeout()` with a tiny small amount of milliseconds works to give a break and let the existing chat messages to be loaded before, then scroll to bottom is successful when chat room page loads first time. Not sure if this would be the best accurate but does the trick. I was researching for some callback function to use in `$bindTo()` which could do a better job than `.then()`, something like `.success()`, `.loaded()`, `.ready()`, something that actually detects and executes when the existing chat messages are already loaded, but could't find any.